requirements.txt
supabase==2.7.4
python-dotenv==1.0.1
requests==2.32.3
feedparser==6.0.11
Pillow==10.4.0
tweepy==4.14.0
